% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_statistic_wrapper.R
\name{auto_statistic_function}
\alias{auto_statistic_function}
\title{Define a function that computes statistics without the need to provide linearized variables.}
\usage{
auto_statistic_function(fn, arg_type)
}
\arguments{
\item{fn}{A function describing the total function to be estimated. All arguments from \code{fn} must be in \code{arg_type}.}

\item{arg_type}{A list that specifies the arguments for the created function, 
including \code{data}, \code{weight}, and optionally \code{param}.}
}
\value{
A function that computes the estimated totals, applies 
the \code{fn} function, and returns a list with two elements:
\describe{
  \item{point}{The point estimate, as a numeric value.}
  \item{lin}{The linearized variable, defined as the dot product between the gradient 
  and the data.}
}
}
\description{
Define a function that computes statistics without the need to provide linearized variables.
The estimated linearized variables are computed using autodifferentiation based
on \code{torch}.
}
\details{
\code{fn} describes the function applied to the estimated totals. This function takes weighted totals as input, 
so weights do not need to be provided in \code{fn}.
}
\examples{
ratio_autolin <- define_statistic_wrapper(
  statistic_function = auto_statistic_function(function(x, y) {return(x / y)}, arg_type),
  arg_type = arg_type
)
}
