% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_linearization_wrapper.R
\name{define_linearization_wrapper}
\alias{define_linearization_wrapper}
\title{TODO}
\usage{
define_linearization_wrapper(linearization_function, allow_factor = FALSE,
  arg_type = list(data = "y", weight = "weight", param = NULL),
  arg_not_affected_by_domain = NULL,
  display_function = standard_display_function)
}
\description{
TODO
}
