% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{define_simple_wrapper}
\alias{define_simple_wrapper}
\title{Define a simple variance wrapper}
\usage{
define_simple_wrapper()
}
\description{
define_simple_wrapper is a convenience wrapper
for define_variance_wrapper for the simplest yet frequent cases: 
\itemize{\item (stratified) simple random sampling \item non-response
correction (if any) through reweighting in Homogeneous response groups 
(HRG) \item calibration (if any)}
It performs technical and statistical checks and (eventually) produces
a ready-to-use variance wrapper adapted for such cases.
}
