% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_statistic_wrapper.R
\name{define_statistic_wrapper}
\alias{define_statistic_wrapper}
\title{Define a statistic wrapper}
\usage{
define_statistic_wrapper(statistic_function, arg_type,
  arg_not_affected_by_domain = NULL,
  display_function = standard_display)
}
\arguments{
\item{statistic_function}{An R function specific to the statistic to
calculate. It should produce at least the point estimator and the
linearized variable associated with the statistic (see Details).}

\item{arg_type}{A named list with three character vectors describing 
the type of each argument of \code{statistic_function} (see Details).}

\item{arg_not_affected_by_domain}{A character vector indicating the (data) 
arguments which should not be affected by domain-splitting. Such parameters
may appear in some complex linearization formula, for instance when the 
At-Risk of Poverty Rate (ARPR) is estimated by region but with a poverty 
line calculated at the national level.}

\item{display_function}{An R function which produces, for each variance 
estimation, the data.frame row to be displayed by the variance estimation 
wrapper. The default display function (\code{standard_display}) uses 
standard metadata to display usual variance indicator (point estimate, 
variance, standard  deviation, coefficient of variation, confidence interval) 
broken down by statistic wrapper, domain (if any) and level (if the variable 
is a factor).}
}
\value{
A function to be used within a variance estimation wrapper to estimate
  a specific statistic (see examples). Its formals are the ones of 
  \code{statistic_function} with the addition of \code{by} and \code{where} 
  (for domain estimation, set to \code{NULL} by default).
}
\description{
\code{define_statistic_wrapper} defines 
  statistic \emph{wrappers} to be used together with
  \code{\link[=define_variance_wrapper]{variance estimation wrappers}}. 
  A statistic wrapper produces both the point estimator and the 
  linearized variable associated with a given statistic to estimate 
  variance on (Deville, 1999). \code{define_statistic_wrapper} is intended 
  for \strong{advanced use only}, standard statistic wrappers are included 
  in the gustave package (see \code{\link[=standard_statistic_wrapper]{standard statistic wrappers}})
}
\details{
When the statistic to estimate is not a total, the application of 
  analytical variance estimation formulae developed for the estimator of a total 
  is not straightforward (Deville, 1999). An asymptotically unbiased variance 
  estimator can nonetheless be obtained if the estimation of variance is performed
  on a variable obtained from the original data through a linearization step. 
  
  \code{define_statistic_wrapper} is the function used to create, for a 
  given statistic, an easy-to-use function which calculates both the point
  estimator and the linearized variable associated with the statistic. These
  operations are implemented by the \code{statistic_function}, which can have
  any needed input (for example \code{num} and \code{denom} for a ratio
  estimator) and should output a list with three named elements: \itemize{
  \item \code{point}: the point estimator of the statistic
  \item \code{lin}: the linearized variable to be passed on to the variance
  estimation formula. If several variables are to be associated with
  the statistics, \code{lin} can be a list itself.
  \item \code{metadata}: optional metadata to be used when displaying
  the results of the variance estimation.
  }
  
  \code{arg_type} is a named list of three elements that describe the nature 
  of the argument of \code{statistic_function}: \itemize{
  \item \code{data}: data argument(s), numerical vector(s) to be used 
  to calculate the point estimate and the linearized variable associated
  with the statistic 
  \item \code{weight}: weight argument, numerical vector to be used 
  as row weights in the linearization formula 
  \item \code{param}: parameters, non-data arguments to be used to 
  control some aspect of the computation of ths statistic}
  
  Statistic wrappers are quite flexible tools to apply a variance function 
  to an estimator requiring a linearization step (e.g. all estimators except 
  the estimator of a total) with virtually no  additional complexity for the
  end-user. To some extent, statistic wrappers can be seen as ggplot2
  \code{geom_} and \code{stat_} functions: they help the end-user in writing 
  down what he or she wants without having to go too  deep into the details 
  of the corresponding layers. 
  
  \code{\link[=standard_statistic_wrapper]{standard statistic wrappers}} 
  are included within the gustave package and automatically added 
  to the variance estimation wrappers. New statistic wrappers can be defined
  using the \code{define_statistic_wrapper} and then explicitly added to the 
  variance estimation wrappers using the \code{objects_to_include} argument.
}
\examples{
### Example from the Information and communication technologies (ICT) survey

# The subset of the (simulated) ICT survey has the following features: 
# - stratified one-stage sampling design of 650 firms;
# - 612 responding firms, non-response correction through reweighting 
# in homogeneous response groups based on economic sub-sector and turnover;
# - calibration on margins (number of firms and turnover broken down
# by economic sub-sector).

# Step 1 : Dummy variance wrapper
# Note : see define_variance_wrapper() for a more 
# realistic variance function and examples.
variance_wrapper <- define_variance_wrapper(
  variance_function = function(y) abs(colSums(y)), 
  reference_id = ict_survey$firm_id, 
  reference_weight = ict_survey$w_calib, 
  default_id = "firm_id"
)
variance_wrapper(ict_survey, total(speed_quanti))

# Step 2 : Redefine the mean statistic wrapper
# The mean() statistic wrapper defined in the gustave 
# package is bulit on top of the ratio() statistic wrapper.
variance_wrapper(ict_survey, mean(speed_quanti))

# Let's redefine it directly from the formula found for instance
# in (Caron, Deville, Sautory, 1998) and without handling NA
# values
mean2 <- define_statistic_wrapper(
  statistic_function = function(y, weight){
    point <- sum(y * weight) / sum(weight)
    lin <- (y - point) / sum(weight)
    list(point = point, lin = lin, metadata = list(n = length(y)))
  },
  arg_type = list(data = "y", weight = "weight")
)
variance_wrapper(ict_survey, mean(speed_quanti), mean2(speed_quanti))

}
\references{
Deville J.-C. (1999), "Variance estimation for complex statistics and 
  estimators: linearization and residual techniques", \emph{Survey Methodology}, 
  25:193–203
}
\seealso{
\code{\link[=standard_statistic_wrapper]{standard statistic wrappers}} 
  \code{\link{define_variance_wrapper}}
}
\author{
Martin Chevalier
}
