% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variance_wrapper.R
\name{define_variance_wrapper}
\alias{define_variance_wrapper}
\title{Create a variance estimation wrapper}
\usage{
define_variance_wrapper(variance_function = NULL, data_arg_name = "y",
  objects_to_include = NULL, default_id = NULL, reference_id = NULL,
  reference_weight = NULL, default_stat = "total", default_alpha = 0.05,
  envir = parent.frame())
}
\description{
Create a variance estimation wrapper
}
