% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_wrapper.R
\name{describe_object}
\alias{describe_object}
\title{Describe a single R object for the summary}
\usage{
describe_object(obj, name)
}
\description{
Describe a single R object for the summary
}
\keyword{internal}
