% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ict_pop}
\alias{ict_pop}
\title{Sampling frame of the Information and communication technologies (ICT)
survey}
\format{
A data frame with 7670 observations and 5 variables:
\describe{
  \item{firm_id}{identifier of the firm}
  \item{division}{identifier of the economic sub-sector}
  \item{employees}{number of employees}
  \item{turnover}{firm turnover, in thousand euros}
  \item{strata}{stratification variable}
}
}
\usage{
ict_pop
}
\description{
A (simulated) dataset containing basic identification information and
auxiliary variables for the sampling of the Information and communication
technologies (ICT) survey in the information and communication sector 
(NACE rev 2 J section).
}
\seealso{
\code{\link{qvar}}, \code{\link{ict_sample}}, \code{\link{ict_survey}}
}
\keyword{datasets}
