% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_samp_area}
\alias{lfs_samp_area}
\title{Sample of areas in the Labour force survey}
\format{A data frame with 4 observations and 3 variables:
\describe{
  \item{id_area}{identifier of the area}
  \item{income}{total annual income of the area in thousand euros (from income registry)}
  \item{pik_area}{first-order inclusion probability of the area (proportional to the number of dwellings in the area)}
}}
\usage{
lfs_samp_area
}
\description{
A (simulated) dataset containing information about 4 geographical 
areas (about 120 dwellings each) sampled for the labour force survey.
}
\seealso{
\code{\link{define_variance_wrapper}}, \code{\link{lfs_samp_dwel}}, \code{\link{lfs_samp_ind}}
}
\keyword{datasets}
