% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_samp_dwel}
\alias{lfs_samp_dwel}
\title{Sample of dwellings in the Labour force survey}
\format{A data frame with 80 observations and 6 variables:
\describe{
  \item{id_dwel}{identifier of the dwelling}
  \item{id_area}{identifier of the area}
  \item{income}{total annual income of the dwelling in thousand euros (from income registry)}
  \item{pik_area}{first-order inclusion probability of the area (proporional to the number of dwellings)}
  \item{pik_dwel}{first-order inclusion probability of the dwelling within the area (20 dwelling sampled per area)}
  \item{pik}{first-order inclusion probability of the dwelling)}
}}
\usage{
lfs_samp_dwel
}
\description{
A (simulated) dataset containing information about 80 dwellings
sampled for the Labour force survey (in the 4 areas described
in \code{\link{lfs_samp_area}}).
}
\seealso{
\code{\link{define_variance_wrapper}} \code{\link{lfs_samp_area}} \code{\link{lfs_samp_ind}}
}
\keyword{datasets}
