% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_samp_ind}
\alias{lfs_samp_ind}
\title{Sample of individuals in the Labour force survey}
\format{
A data frame with 157 observations and 5 variables:
\describe{
  \item{id_ind}{identifier of the individual}
  \item{id_dwel}{identifier of the dwelling}
  \item{income}{total annual income of the individual in thousand euros (from income registry)}
  \item{unemp}{unemployment status}
  \item{sampling_weight}{sampling weight of the individual (inverse of the first-order inclusion probability of the dwelling)}
}
}
\usage{
lfs_samp_ind
}
\description{
A (simulated) dataset containing information about 157 individuals
sampled for the Labour force survey (all members of the 80 dwellings
described in \code{\link{lfs_samp_dwel}}). It also contains the 
unemployment status extracted from the survey file (no non-response).
}
\seealso{
\code{\link{define_variance_wrapper}}, \code{\link{lfs_samp_area}}, \code{\link{lfs_samp_dwel}}
}
\keyword{datasets}
