% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearization_wrapper_laeken.R
\name{linearization_wrapper_laeken}
\alias{linearization_wrapper_laeken}
\alias{arpr}
\alias{gini}
\alias{arpr}
\alias{gini}
\title{Linearization wrappers for Laeken indicators}
\usage{
arpr(y, percentage = 60, order_quant = 50L, by = NULL, where = NULL,
  ...)

gini(y, by = NULL, where = NULL, ...)
}
\description{
Linearization wrappers for Laeken indicators
}
