% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearization_wrapper_standard.R
\name{linearization_wrapper_standard}
\alias{linearization_wrapper_standard}
\alias{total}
\alias{ratio}
\alias{mean}
\alias{diffratio}
\alias{ratio}
\alias{mean}
\alias{diffratio}
\title{Standard linearization wrappers}
\usage{
total(y, weight, by = NULL, where = NULL, technical_arg)

ratio(num, denom, weight, by = NULL, where = NULL, technical_arg)

mean(y, weight, by = NULL, where = NULL, technical_arg)

diffratio(num1, denom1, num2, denom2, weight, by = NULL, where = NULL,
  technical_arg)
}
\description{
Standard linearization wrappers
}
