% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbs_pop}
\alias{sbs_pop}
\title{Sampling frame of the Structural business statistics (SBS)
survey in the construction sector (NACE rev 2 F section)}
\format{A data frame with 4950 observations and 4 variables:
\describe{
  \item{id}{identifier of the firm}
  \item{division}{identifier of the economic sub-sector}
  \item{label}{label of the economic sub-sector}
  \item{turnover}{firm turnover, in euros}
}}
\usage{
sbs_pop
}
\description{
A (fake) dataset containing basic identification information and
auxiliary variables for the sampling of the Structural business
statistics (SBS) survey in the construction sector (NACE rev 2 F section).
}
\keyword{datasets}
