% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbs_sample}
\alias{sbs_sample}
\title{Sample of the Structural business statistics (SBS)
survey in the construction sector (NACE rev 2 F section)}
\format{A data frame with 308 observations and 8 variables:
\describe{
  \item{id}{identifier of the firm}
  \item{division}{identifier of the economic sub-sector}
  \item{label}{label of the economic sub-sector}
  \item{turnover}{firm turnover, in euros}
  \item{w_sample}{sampling weight}
  \item{resp}{boolean indicating whether the firm did respond to the survey or not}
  \item{hrg}{homogeneous response group used for the unit non-response correction}
  \item{w_nr}{weight after unit non-response correction}
}}
\usage{
sbs_sample
}
\description{
A (fake) dataset containing sampling information about the sample 
of the Structural business statistics (SBS) survey in the construction 
sector (NACE rev 2 F section).
}
\keyword{datasets}
