% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbs_survey}
\alias{sbs_survey}
\title{Survey data of the Structural business statistics (SBS)
survey in the construction sector (NACE rev 2 F section)}
\format{A data frame with 267 observations and 12 variables:
\describe{
  \item{id}{identifier of the firm}
  \item{division}{identifier of the economic sub-sector}
  \item{label}{label of the economic sub-sector}
  \item{turnover}{firm turnover, in euros}
  \item{investment}{firm investment, in euros}
  \item{N_41, N_42, N_43, turnover_41, turnover_42, turnover_43}{calibration variables (by economic sub-sector)}
  \item{w_calib}{calibrated weight}
}}
\usage{
sbs_survey
}
\description{
A (fake) dataset containing calibration and survey variables of the respondents 
to the Structural business statistics (SBS) survey in the construction 
sector (NACE rev 2 F section).
}
\keyword{datasets}
