% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfun.R
\name{varGR}
\alias{varGR}
\title{TODO}
\usage{
varGR(y, alphakl, pikl = matrix(1, ncol = NCOL(alphakl), nrow =
  NROW(alphakl)), g = rep(0, NROW(alphakl)))
}
\description{
TODO
}
\references{
doc de travail M2015 3 Gros Moussallam p.11. et p. 19 pour la version matricielle
}
