% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_function.R
\name{varSYG}
\alias{varSYG}
\title{Sen-Yates-Grundy variance estimator}
\usage{
varSYG(y = NULL, pikl, precalc = NULL)
}
\description{
Sen-Yates-Grundy variance estimator
}
\references{
doc de travail M2015 3 Gros Moussallam p. 19
}
