% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varwrap_factory.R
\name{varwrap_factory}
\alias{varwrap_factory}
\title{Create a variance estimation wrapper}
\usage{
varwrap_factory(varfun = NULL, objects_to_include = NULL,
  id_default = NULL, idref = NULL, wref = NULL, data_arg = "y",
  envir = parent.frame())
}
\description{
Create a variance estimation wrapper
}
