% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_linearization_wrapper.R
\name{define_linearization_wrapper}
\alias{define_linearization_wrapper}
\title{TODO}
\usage{
define_linearization_wrapper(linearization_function, weight_arg_name = "w",
  allow_factor = FALSE, display_function = standard_display_function)
}
\description{
TODO
}
