% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variance_wrapper.R
\name{define_variance_wrapper}
\alias{define_variance_wrapper}
\title{Define a variance estimation wrapper}
\usage{
define_variance_wrapper(variance_function, arg_type = NULL,
  reference_id = NULL, technical_data = NULL, default = list(id = NULL,
  weight = NULL, stat = "total", alpha = 0.05), objects_to_include = NULL,
  objects_to_include_from = parent.frame())
}
\arguments{
\item{variance_function}{An R function, with input a data matrix and possibly 
other arguments (see \code{arg_type}) and output a numeric vector of estimated 
variances (or a list whose first element is a numeric vector of estimated variances).}

\item{arg_type}{A named list with three character vectors describing 
the type of each argument of \code{variance_function}: \itemize{
\item \code{data}: data argument (only one allowed), the numerical matrix of 
variables of interest to be used in the variance estimation formula 
\item \code{tech}: technical data arguments, information to be used within 
the variance estimation function 
\item \code{param}: parameters, non-data arguments to be used to control 
some aspect of the variance estimation} 
If \code{NULL}, the first argument of \code{variance_function} is considered 
as the data argument and all others as technical data arguments (no parameters 
by default).}

\item{reference_id}{A vector containing the ids of all the responding units
of the survey. It is compared with \code{default$id} (see below) to check
whether some observations are missing in the survey file. Observations are
reordered according to \code{reference_id} prior to the variance estimation.}

\item{technical_data}{A named list of all technical data needed to perform the variance
estimation. It must be consistent with \code{variance_function} arguments described as 
technical data arguments (\code{tech}) in \code{arg_type}.}

\item{default}{A named list specifying the default values for: \itemize{
  \item \code{id}: the name of the default identifying variable in the survey 
  file. It can also be an unevaluated expression (enclosed in \code{substitute()}) to be 
  evaluated within the survey file.
  \item \code{weight}: the name of the default weight variable in the survey file. 
  It can also be an unevaluated expression (enclosed in \code{substitute()}) to be 
  evaluated within the survey file.
  \item \code{stat}: the name of the default statistic to compute when none is specified. 
  It is set to \code{"total"} by default.
  \item \code{alpha}: the default threshold for confidence interval derivation. 
  It is set to \code{0.05} by default.
}}

\item{objects_to_include}{A character vector indicating the name of 
additional R objects to include within the variance wrapper.}

\item{objects_to_include_from}{The environment to which the additional R 
objects belong.}
}
\value{
An R function that makes the estimation of variance based on the provided 
variance function easier. Its parameters are:
  \itemize{
   \item \code{data}: the survey data where the interest variables are stored
   \item \code{...}: one or more calls to a linearization wrapper (see examples
   and \code{\link[=linearization_wrapper_standard]{standard linearization wrappers}})
   \item \code{where}: a logical vector indicating a domain on which the variance
   estimation is conducted
   \item \code{by}: a qualitative variable whose levels are used to define domains
   on which the variance estimation is conducted
   \item \code{stat}: a character vector of size 1 indicating the linearization
   wrapper to use when none is specified. Its default value depends on
   the value of \code{default_stat} in \code{define_variance_wrapper}
   \item \code{alpha}: a numeric vector of size 1 indicating the threshold
   for confidence interval derivation. Its default value depends on
   the value of \code{default_alpha} in \code{define_variance_wrapper}
   \item \code{id}: a character vector of size 1 containing the name of
   the identifying variable in the survey file. It can also be an 
   unevaluated expression (using \code{substitute()}) to be evaluated within
   the survey file. Its default value depends on the value of 
   \code{default_id} in \code{define_variance_wrapper}
   \item \code{envir}: an environment containing a binding to \code{data}
 }
}
\description{
Given a variance estimation \emph{function} (specific to a 
  survey), \code{define_variance_wrapper} defines a variance estimation 
  \emph{wrapper} easier to use (e.g. automatic domain estimation, 
  linearization).
}
\details{
Defining variance estimation wrappers is the \strong{key feature} of
  the \code{gustave} package.
  
  Analytical variance estimation is often difficult to carry out by 
  non-specialists owing to the complexity of the underlying sampling 
  and estimation methodology. This complexity yields complex \emph{variance estimation 
  functions} which are most often only used by the sampling expert who 
  actually wrote them. A \emph{variance estimation wrapper} is an 
  intermediate function that is "wrapped around" the (complex) variance 
  estimation function in order to provide the non-specialist with 
  user-friendly features: \itemize{ \item checks for consistency between the 
  provided dataset and the survey characteristics \item factor discretization
  \item domain estimation \item linearization of complex statistics (see 
  \code{\link[=linearization_wrapper_standard]{standard linearization wrappers}})}
  
  \code{define_variance_wrapper} allows the sampling expert to define a 
  variance estimation wrapper around a given variance estimation function and
  set its default parameters. The produced variance estimation wrapper will 
  be stand-alone in the sense that it can contain all technical data necessary
  to carry out the estimation (see \code{technical_data}).
}
\examples{
### Example from the Information and communication technologies (ICT) survey

# The subset of the (simulated) ICT survey has the following features: 
# - stratified one-stage sampling design of 650 firms;
# - 612 responding firms, non-response correction through reweighting 
# in homogeneous response groups based on economic sub-sector and turnover;
# - calibration on margins (number of firms and turnover broken down
# by economic sub-sector).

# Step 1 : Definition of a variance function

variance_function <- function(y, x, w, samp){
  
  # Calibration
  y <- rescal(y, x = x, w = w)
  
  # Non-response
  y <- add0(y, rownames = samp$firm_id)
  var_nr <- var_pois(y, pik = samp$response_prob_est, w = samp$w_sample)
  
  # Sampling
  y <- y / samp$response_prob_est
  var_sampling <- var_srs(y, pik = 1 / samp$w_sample, strata = samp$division)
  
  var_sampling + var_nr
  
}

# y is the matrix of variables of interest, x, w, and samp are some technical data:
technical_data <- list(
  
  # x: calibration variables matrix
  x = as.matrix(ict_survey[
    order(ict_survey$firm_id),
    c(paste0("N_", 58:63), paste0("turnover_", 58:63))
  ]),

  # w: calibrated weights
  w = ict_survey$w_calib[order(ict_survey$firm_id)],
  
  # samp: sample file
  samp = ict_sample
  
)

# Test of the variance function
y <- matrix(ict_survey$speed_quanti, dimnames = list(ict_survey$firm_id))
with(technical_data, variance_function(y, samp = samp, x = x, w = w))

# Step 2 : Definition of a variance wrapper

variance_wrapper <- define_variance_wrapper(
  variance_function = variance_function,
  reference_id = ict_survey$firm_id, 
  technical_data = technical_data,
  default = list(id = "firm_id", weight = "w_calib")
)

# The object "technical_data" is embedded within the function 
# variance_wrapper together with variance_function
ls(environment(variance_wrapper))
ls(environment(variance_wrapper)$technical_data)
# Note : variance_wrapper is a closure
# (http://adv-r.had.co.nz/Functional-programming.html#closures)
# As a consequence, the variance wrapper will work even if 
# these objects are removed from globalenv()
rm(x, w, samp)

# Step 3 : Features of the variance wrapper

# Better display of results
variance_wrapper(ict_survey, speed_quanti)

# Mean linearization
variance_wrapper(ict_survey, mean(speed_quanti))
# Ratio linearization
variance_wrapper(ict_survey, ratio(turnover, employees))

# Discretization of qualitative variables
variance_wrapper(ict_survey, speed_quali)
# On-the-fly recoding
variance_wrapper(ict_survey, speed_quali == "Between 2 and 10 Mbs")

# 1-domain estimation
variance_wrapper(ict_survey, speed_quanti, where = division == "58")
# Multiple domains estimation
variance_wrapper(ict_survey, speed_quanti, by = division)

# Multiple variables at a time
variance_wrapper(ict_survey, speed_quanti, big_data)
variance_wrapper(ict_survey, speed_quanti, mean(big_data))
# Flexible syntax for where and by arguments
# (similar to the aes() function in ggplot2)
variance_wrapper(ict_survey, where = division == "58", 
  mean(speed_quanti), mean(big_data * 100)
)
variance_wrapper(ict_survey, where = division == "58", 
  mean(speed_quanti), mean(big_data * 100, where = division == "61")
)
variance_wrapper(ict_survey, where = division == "58", 
  mean(speed_quanti), mean(big_data * 100, where = NULL)
)

}
\seealso{
\code{\link[=linearization_wrapper_standard]{standard linearization wrappers}} \code{\link{varDT}}
}
\author{
Martin Chevalier
}
