% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everest.R
\name{everest}
\alias{everest}
\title{Perform a variance estimation for one-stage stratified surveys 
taking out-of-scope units, non-response and calibration into acount}
\usage{
everest(data, ..., by = NULL, where = NULL, alpha = 0.05,
  display = TRUE, id, dissemination_dummy, dissemination_weight,
  sampling_weight, strata = NULL, scope_dummy = NULL, nrc_weight = NULL,
  response_dummy = NULL, nrc_dummy = NULL, calibration_weight = NULL,
  calibration_dummy = NULL, calibration_var = NULL, define = FALSE,
  envir = parent.frame())
}
\arguments{
\item{data}{The \code{data.frame} containing all the technical information
required to prepare the variance estimation process (see other arguments 
below). Note that this file should contain all the units sampled, 
including the out-of-scope and non-responding units. If a variance
estimation is to be performed right away (when \code{define = FALSE}),
it should also contain the variables of interest.}

\item{...}{One or more calls to a statistic wrapper (e.g. \code{total()}, 
\code{mean()}, \code{ratio()}). See examples and 
\code{\link[=standard_statistic_wrapper]{standard statistic wrappers}})}

\item{by}{A qualitative variable whose levels are used to define domains
on which the variance estimation is performed.}

\item{where}{A logical vector indicating a domain on which the variance 
estimation is to be performed.}

\item{alpha}{A numeric vector of length 1 indicating the threshold
for confidence interval derivation (\code{0.05} by default).}

\item{display}{A logical verctor of length 1 indicating whether
the result of the estimation should be displayed or not.}

\item{id}{The identification variable of the units in \code{data}. 
It should be unique for each row in \code{data} and not contain any 
missing values.}

\item{dissemination_dummy}{A character vector of length 1, the name
of the logical variable in \code{data} indicating whether the unit
does appear in the disseminated file and should be used for point
estimates.}

\item{dissemination_weight}{A character vector of length 1, the name
of the numerical variable in \code{data} corresponding to the 
dissemination weight of the survey. It should not contain any missing 
(NA) values.}

\item{sampling_weight}{A character vector of length 1, the name of the 
numeric variable in \code{data} corresponding to the sampling weights 
of the survey. It should not contain any missing (NA) values.}

\item{strata}{A character vector of length 1, the name of the factor 
variable in \code{data} whose level match the stratification
used in the survey. Character variables are coerced to factor.
If defined, it should not contain any missing value. If \code{NULL},
the variance estimation process does not take any stratification
into account.}

\item{scope_dummy}{A character vector of length 1, the name of the logical 
variable in \code{data} indicating whether the unit belongs to the
scope of the survey or not. Numerical variables are coerced to logical.
If defined, it should not contain any missing value. If \code{NULL},
all units are supposed to be within the scope of the survey.}

\item{nrc_weight}{A character vector of length 1, the name of the 
numerical variable in \code{data} corresponding to the weights
after non-response correction. If defined, all responding units 
(see below) should have a non-missing value. If \code{NULL}, all
units are supposed to be responding and the variance estimation
process does not include a second phase in order to take non-response
into account.}

\item{response_dummy}{A character vector of length 1, the name of of the logical 
variable in \code{data} indicating whether the unit is a responding 
unit or not. Numerical variables are coerced to logical. \code{response_dummy}
should be defined as long as a \code{nrc_weight} is provided. All units 
in the scope of the survey should have a non-missing value.}

\item{nrc_dummy}{A character vector of length 1, the name of
the logical variable in \code{data} indicating whether the
units did take part in the non-response correction process. 
All units in the scope of the survey should have a non-missing 
value.}

\item{calibration_weight}{A character vector of length 1, the name of the 
numerical variable in \code{data} corresponding to the calibrated
weights. If defined, all responding units (see above) should have 
a non-missing value. If \code{NULL}, the variance estimation
process does not take any calibration step into account.}

\item{calibration_dummy}{A character vector of length 1, the name of of the logical 
variable in \code{data} indicating whether the unit did take part
in the calibration process or not. Numerical variables are coerced to 
logical. If defined, all responding units should have a non-missing
value. If \code{NULL}, calibration is supposed to have been performed
on all responding units.}

\item{calibration_var}{A character vector, the name of the variable(s) used in
the calibration process. Logical variables are coerced to numeric. 
Character and factor variables are automatically discretized. 
\code{calibration_var} should be defined as long as a \code{calibration_weight} is 
provided. All units taking part in the calibration process should have
only non-missing values for all variables in \code{calibration_var}.}

\item{define}{Logical vector of lentgh 1. Should a variance wrapper
be defined instead of performing a variance estimation?}

\item{envir}{An environment containing a binding to \code{data}.}
}
\description{
\code{everest} is a function performing analytical 
variance estimation in most common cases, that is: \itemize{\item stratified 
simple random sampling \item correct handling of out-of-scope units \item 
non-response correction (if any) through reweighting \item calibration (if any)}

Used with \code{define = TRUE}, it defines a so-called variance wrapper, that 
is a standalone ready-to-use function that can be applied to the survey dataset 
without having to specify the methodological characteristics of the survey.
}
\details{
\code{everest} aims at being a straightforward variance
  estimation tool in the most common cases in social or business
  surveys, that is a stratified simple random sampling with 
  additional reweighting steps to correct non-response bias
  or reduce variance through calibration.
  
  It performs not only tehnical but also methodological checks
  in order to ensure that the variance estimation methodology 
  do apply (e.g. equal probability of inclusion within strata, 
  number of units per stratum). 
  
  When the sampling design or reweighting schemes are more complex,
  a specific variance estimation wrapper may be defined using
  the \code{\link{define_variance_wrapper}} function.
  
  "everest" strands for (in French) "Variance estimation
  for stratified simple random sampling surveys with reweighting".
}
\seealso{
\code{\link{define_variance_wrapper}} \code{\link{standard_statistic_wrapper}}
}
