% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{everest}
\alias{everest}
\alias{define_simple_wrapper}
\alias{everest}
\alias{define_simple_wrapper}
\title{Perform a variance estimation in most common cases}
\usage{
everest(data, ..., by = NULL, where = NULL, id = NULL, alpha = 0.05,
  display = TRUE, envir = parent.frame(), samp_weight, strata = NULL,
  scope_dummy = NULL, no_reweighting_dummy = NULL, nrc_weight = NULL,
  resp_dummy = NULL, calib_weight = NULL, calib_dummy = NULL,
  calib_var = NULL, define = FALSE)

define_simple_wrapper(data, id, samp_weight, strata = NULL,
  scope_dummy = NULL, no_reweighting_dummy = NULL, nrc_weight = NULL,
  resp_dummy = NULL, calib_weight = NULL, calib_dummy = NULL,
  calib_var = NULL)
}
\arguments{
\item{data}{The \code{data.frame} containing all the technical information
required to prepare the variance estimation process (see other arguments 
below). Note that this file should contain all the units sampled, 
including the out-of-scope and non-responding units. For the \code{everest}
function, it should also contain the variables of interest (the variables
to perform the variance estimation on).}

\item{...}{One or more calls to a linearization wrapper (e.g. \code{total()}, 
\code{mean()}, \code{ratio()}). See examples and 
\code{\link[=linearization_wrapper_standard]{standard linearization wrappers}})}

\item{by}{A qualitative variable whose levels are used to define domains
on which the variance estimation is performed.}

\item{where}{A logical vector indicating a domain on which the variance 
estimation is to be performed.}

\item{id}{The identification variable of the units in \code{data}. 
It should be unique for each row in \code{data} and not contain any 
missing values.}

\item{alpha}{A numeric vector of length 1 indicating the threshold
for confidence interval derivation (\code{0.05} by default).}

\item{display}{A logical verctor of length 1 indicating whether
the result of the estimation should be displayed or not.}

\item{envir}{An environment containing a binding to \code{data}.}

\item{samp_weight}{A character vector of length 1, the name of the 
numeric variable in \code{data} corresponding to the sampling weights 
of the survey. It should not contain any missing values.}

\item{strata}{A character vector of length 1, the name of the factor 
variable in \code{data} whose level match the stratification
used in the survey. Character variables are coerced to factor.
If defined, it should not contain any missing value. If \code{NULL},
the variance estimation process does not take any stratification
into account.}

\item{scope_dummy}{A character vector of length 1, the name of the logical 
variable in \code{data} indicating whether the unit belongs to the
scope of the survey or not. Numerical variables are coerced to logical.
If defined, it should not contain any missing value. If \code{NULL},
all units are supposed to be within the scope of the survey.}

\item{no_reweighting_dummy}{A character vector of length 1, the name of
the logical variable in \code{data} indicating whether the units
did take part in the reweighting process (non-response correction, 
calibration). Units not taking part in the reweighting process
are often atypical units whose weight should not be inflated.}

\item{nrc_weight}{A character vector of length 1, the name of the 
numerical variable in \code{data} corresponding to the weights
after non-response correction. If defined, all responding units 
(see below) should have a non-missing value. If \code{NULL}, all
units are supposed to be responding and the variance estimation
process does not include a second phase in order to take non-response
into account.}

\item{resp_dummy}{A character vector of length 1, the name of of the logical 
variable in \code{data} indicating whether the unit is a responding 
unit or not. Numerical variables are coerced to logical. \code{resp_dummy}
should be defined as long as a \code{nrc_weight} is provided. All units 
in the scope of the survey should have a non-missing value.}

\item{calib_weight}{A character vector of length 1, the name of the 
numerical variable in \code{data} corresponding to the calibrated
weights. If defined, all responding units (see above) should have 
a non-missing value. If \code{NULL}, the variance estimation
process does not take any calibration step into account.}

\item{calib_dummy}{A character vector of length 1, the name of of the logical 
variable in \code{data} indicating whether the unit did take part
in the calibration process or not. Numerical variables are coerced to 
logical. If defined, all responding units should have a non-missing
value. If \code{NULL}, calibration is supposed to have been performed
on all responding units.}

\item{calib_var}{A character vector, the name of the variable(s) used in
the calibration process. Logical variables are coerced to numeric. 
Character and factor variables are automatically discretized. 
\code{calib_var} should be defined as long as a \code{calib_weight} is 
provided. All units taking part in the calibration process should have
only non-missing values for all variables in \code{calib_var}.}

\item{define}{Logical vector of lentgh 1. Should a variance wrapper
be defined instead of performing a variance estimation?}
}
\description{
\code{everest} is a ready-to-use function performing
analytical variance estimation in most common cases, that is: 
\itemize{\item (stratified) simple random sampling \item non-response
correction (if any) through reweighting \item calibration (if any)}

\code{define_simple_wrapper} (or \code{everest} with \code{define = TRUE})
defines a so-called variance wrapper, that is a standalone function that
can be applied to the survey dataset in order to perform the variance estimation
without having to specify the methodological characteristics of the survey.
}
