% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ict_survey}
\alias{ict_survey}
\title{Survey data of the Information and communication technologies (ICT)
survey}
\format{
A data frame with 612 observations and 11 variables:
\describe{
  \item{firm_id}{identifier of the firm}
  \item{division}{identifier of the economic sub-sector}
  \item{employees}{number of employees}
  \item{turnover}{firm turnover, in euros}
  \item{w_calib}{calibrated weight}
  \item{speed_quanti, speed_quanti_NA}{internet connection speed of the firm in Mbps, without or with missing values}
  \item{speed_quali, speed_quali_NA}{internet connection speed of the firm recoded in classes, without or with missing values}
  \item{big_data, big_data_NA}{use of big data analytics within the firm, without or with missing values}
}
}
\usage{
ict_survey
}
\description{
A (simulated) dataset containing calibration and survey variables of the respondents 
to the Information and communication technologies (ICT)
survey in the information and communication sector (NACE rev 2 J section)
}
\seealso{
\code{\link{qvar}}, \code{\link{ict_pop}}, \code{\link{ict_sample}}
}
\keyword{datasets}
