% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearization_wrapper_standard.R
\name{linearization_wrapper_standard}
\alias{linearization_wrapper_standard}
\alias{total}
\alias{ratio}
\alias{mean}
\alias{diffratio}
\alias{ratio}
\alias{mean}
\alias{diffratio}
\title{Standard linearization wrappers}
\usage{
total(y, by = NULL, where = NULL, ...)

ratio(num, denom, by = NULL, where = NULL, ...)

mean(y, by = NULL, where = NULL, ...)

diffratio(num1, denom1, num2, denom2, by = NULL, where = NULL, ...)
}
\description{
Standard linearization wrappers
}
